Struct google_datacatalog1_beta1::api::ProjectMethods

pub struct ProjectMethods<'a, C> where
    C: 'a,  { /* fields omitted */ }

A builder providing access to all methods supported on project resources. It is not used directly, but through the DataCatalog hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate hyper_rustls;
extern crate yup_oauth2 as oauth2;
extern crate google_datacatalog1_beta1 as datacatalog1_beta1;
 
use std::default::Default;
use oauth2;
use datacatalog1_beta1::DataCatalog;
 
let secret: oauth2::ApplicationSecret = Default::default();
let auth = yup_oauth2::InstalledFlowAuthenticator::builder(
        secret,
        yup_oauth2::InstalledFlowReturnMethod::HTTPRedirect,
    ).build().await.unwrap();
let mut hub = DataCatalog::new(hyper::Client::builder().build(hyper_rustls::HttpsConnector::with_native_roots()), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `locations_entry_groups_create(...)`, `locations_entry_groups_delete(...)`, `locations_entry_groups_entries_create(...)`, `locations_entry_groups_entries_delete(...)`, `locations_entry_groups_entries_get(...)`, `locations_entry_groups_entries_get_iam_policy(...)`, `locations_entry_groups_entries_list(...)`, `locations_entry_groups_entries_patch(...)`, `locations_entry_groups_entries_tags_create(...)`, `locations_entry_groups_entries_tags_delete(...)`, `locations_entry_groups_entries_tags_list(...)`, `locations_entry_groups_entries_tags_patch(...)`, `locations_entry_groups_entries_test_iam_permissions(...)`, `locations_entry_groups_get(...)`, `locations_entry_groups_get_iam_policy(...)`, `locations_entry_groups_list(...)`, `locations_entry_groups_patch(...)`, `locations_entry_groups_set_iam_policy(...)`, `locations_entry_groups_tags_create(...)`, `locations_entry_groups_tags_delete(...)`, `locations_entry_groups_tags_list(...)`, `locations_entry_groups_tags_patch(...)`, `locations_entry_groups_test_iam_permissions(...)`, `locations_tag_templates_create(...)`, `locations_tag_templates_delete(...)`, `locations_tag_templates_fields_create(...)`, `locations_tag_templates_fields_delete(...)`, `locations_tag_templates_fields_enum_values_rename(...)`, `locations_tag_templates_fields_patch(...)`, `locations_tag_templates_fields_rename(...)`, `locations_tag_templates_get(...)`, `locations_tag_templates_get_iam_policy(...)`, `locations_tag_templates_patch(...)`, `locations_tag_templates_set_iam_policy(...)`, `locations_tag_templates_test_iam_permissions(...)`, `locations_taxonomies_create(...)`, `locations_taxonomies_delete(...)`, `locations_taxonomies_export(...)`, `locations_taxonomies_get(...)`, `locations_taxonomies_get_iam_policy(...)`, `locations_taxonomies_import(...)`, `locations_taxonomies_list(...)`, `locations_taxonomies_patch(...)`, `locations_taxonomies_policy_tags_create(...)`, `locations_taxonomies_policy_tags_delete(...)`, `locations_taxonomies_policy_tags_get(...)`, `locations_taxonomies_policy_tags_get_iam_policy(...)`, `locations_taxonomies_policy_tags_list(...)`, `locations_taxonomies_policy_tags_patch(...)`, `locations_taxonomies_policy_tags_set_iam_policy(...)`, `locations_taxonomies_policy_tags_test_iam_permissions(...)`, `locations_taxonomies_set_iam_policy(...)` and `locations_taxonomies_test_iam_permissions(...)`
// to build up your call.
let rb = hub.projects();

Implementations

impl<'a, C> ProjectMethods<'a, C>

pub fn locations_entry_groups_entries_tags_create(
&self,
request: GoogleCloudDatacatalogV1beta1Tag,
parent: &str
) -> ProjectLocationEntryGroupEntryTagCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a tag on an Entry. Note: The project identified by the parent parameter for the tag and the tag template used to create the tag must be from the same organization.

Arguments

request - No description provided.

parent - Required. The name of the resource to attach this tag to. Tags can be attached to Entries. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id} Note that this Tag and its child resources may not actually be stored in the location in this name.

pub fn locations_entry_groups_entries_tags_delete(
&self,
name: &str
) -> ProjectLocationEntryGroupEntryTagDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a tag.

Arguments

name - Required. The name of the tag to delete. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id}/tags/{tag_id}

pub fn locations_entry_groups_entries_tags_list(
&self,
parent: &str
) -> ProjectLocationEntryGroupEntryTagListCall<'a, C>

Create a builder to help you perform the following task:

Lists the tags on an Entry.

Arguments

parent - Required. The name of the Data Catalog resource to list the tags of. The resource could be an Entry or an EntryGroup. Examples: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id} * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id}

pub fn locations_entry_groups_entries_tags_patch(
&self,
request: GoogleCloudDatacatalogV1beta1Tag,
name: &str
) -> ProjectLocationEntryGroupEntryTagPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing tag.

Arguments

request - No description provided.

name - The resource name of the tag in URL format. Example: * projects/{project_id}/locations/{location}/entrygroups/{entry_group_id}/entries/{entry_id}/tags/{tag_id} where tag_id is a system-generated identifier. Note that this Tag may not actually be stored in the location in this name.

pub fn locations_entry_groups_entries_create(
&self,
request: GoogleCloudDatacatalogV1beta1Entry,
parent: &str
) -> ProjectLocationEntryGroupEntryCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates an entry. Only entries of 'FILESET' type or user-specified type can be created. Users should enable the Data Catalog API in the project identified by the parent parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information). A maximum of 100,000 entries may be created per entry group.

Arguments

request - No description provided.

parent - Required. The name of the entry group this entry is in. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id} Note that this Entry and its child resources may not actually be stored in the location in this name.

pub fn locations_entry_groups_entries_delete(
&self,
name: &str
) -> ProjectLocationEntryGroupEntryDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an existing entry. Only entries created through CreateEntry method can be deleted. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

name - Required. The name of the entry. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id}

pub fn locations_entry_groups_entries_get(
&self,
name: &str
) -> ProjectLocationEntryGroupEntryGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an entry.

Arguments

name - Required. The name of the entry. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id}

pub fn locations_entry_groups_entries_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationEntryGroupEntryGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. A NOT_FOUND error is returned if the resource does not exist. An empty policy is returned if the resource exists but does not have a policy set on it. Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. Callers must have following Google IAM permission - datacatalog.tagTemplates.getIamPolicy to get policies on tag templates. - datacatalog.entries.getIamPolicy to get policies on entries. - datacatalog.entryGroups.getIamPolicy to get policies on entry groups.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_entry_groups_entries_list(
&self,
parent: &str
) -> ProjectLocationEntryGroupEntryListCall<'a, C>

Create a builder to help you perform the following task:

Lists entries.

Arguments

parent - Required. The name of the entry group that contains the entries, which can be provided in URL format. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}

pub fn locations_entry_groups_entries_patch(
&self,
request: GoogleCloudDatacatalogV1beta1Entry,
name: &str
) -> ProjectLocationEntryGroupEntryPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing entry. Users should enable the Data Catalog API in the project identified by the entry.name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

request - No description provided.

name - Output only. The Data Catalog resource name of the entry in URL format. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id} Note that this Entry and its child resources may not actually be stored in the location in this name.

pub fn locations_entry_groups_entries_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationEntryGroupEntryTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns the caller's permissions on a resource. If the resource does not exist, an empty set of permissions is returned (We don't return a NOT_FOUND error). Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. A caller is not required to have Google IAM permission to make this request.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_entry_groups_tags_create(
&self,
request: GoogleCloudDatacatalogV1beta1Tag,
parent: &str
) -> ProjectLocationEntryGroupTagCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a tag on an Entry. Note: The project identified by the parent parameter for the tag and the tag template used to create the tag must be from the same organization.

Arguments

request - No description provided.

parent - Required. The name of the resource to attach this tag to. Tags can be attached to Entries. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id} Note that this Tag and its child resources may not actually be stored in the location in this name.

pub fn locations_entry_groups_tags_delete(
&self,
name: &str
) -> ProjectLocationEntryGroupTagDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a tag.

Arguments

name - Required. The name of the tag to delete. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id}/tags/{tag_id}

pub fn locations_entry_groups_tags_list(
&self,
parent: &str
) -> ProjectLocationEntryGroupTagListCall<'a, C>

Create a builder to help you perform the following task:

Lists the tags on an Entry.

Arguments

parent - Required. The name of the Data Catalog resource to list the tags of. The resource could be an Entry or an EntryGroup. Examples: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id} * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}/entries/{entry_id}

pub fn locations_entry_groups_tags_patch(
&self,
request: GoogleCloudDatacatalogV1beta1Tag,
name: &str
) -> ProjectLocationEntryGroupTagPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an existing tag.

Arguments

request - No description provided.

name - The resource name of the tag in URL format. Example: * projects/{project_id}/locations/{location}/entrygroups/{entry_group_id}/entries/{entry_id}/tags/{tag_id} where tag_id is a system-generated identifier. Note that this Tag may not actually be stored in the location in this name.

pub fn locations_entry_groups_create(
&self,
request: GoogleCloudDatacatalogV1beta1EntryGroup,
parent: &str
) -> ProjectLocationEntryGroupCreateCall<'a, C>

Create a builder to help you perform the following task:

A maximum of 10,000 entry groups may be created per organization across all locations. Users should enable the Data Catalog API in the project identified by the parent parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

request - No description provided.

parent - Required. The name of the project this entry group is in. Example: * projects/{project_id}/locations/{location} Note that this EntryGroup and its child resources may not actually be stored in the location in this name.

pub fn locations_entry_groups_delete(
&self,
name: &str
) -> ProjectLocationEntryGroupDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes an EntryGroup. Only entry groups that do not contain entries can be deleted. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

name - Required. The name of the entry group. For example, projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}.

pub fn locations_entry_groups_get(
&self,
name: &str
) -> ProjectLocationEntryGroupGetCall<'a, C>

Create a builder to help you perform the following task:

Gets an EntryGroup.

Arguments

name - Required. The name of the entry group. For example, projects/{project_id}/locations/{location}/entryGroups/{entry_group_id}.

pub fn locations_entry_groups_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationEntryGroupGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. A NOT_FOUND error is returned if the resource does not exist. An empty policy is returned if the resource exists but does not have a policy set on it. Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. Callers must have following Google IAM permission - datacatalog.tagTemplates.getIamPolicy to get policies on tag templates. - datacatalog.entries.getIamPolicy to get policies on entries. - datacatalog.entryGroups.getIamPolicy to get policies on entry groups.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_entry_groups_list(
&self,
parent: &str
) -> ProjectLocationEntryGroupListCall<'a, C>

Create a builder to help you perform the following task:

Lists entry groups.

Arguments

parent - Required. The name of the location that contains the entry groups, which can be provided in URL format. Example: * projects/{project_id}/locations/{location}

pub fn locations_entry_groups_patch(
&self,
request: GoogleCloudDatacatalogV1beta1EntryGroup,
name: &str
) -> ProjectLocationEntryGroupPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates an EntryGroup. The user should enable the Data Catalog API in the project identified by the entry_group.name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

request - No description provided.

name - The resource name of the entry group in URL format. Example: * projects/{project_id}/locations/{location}/entryGroups/{entry_group_id} Note that this EntryGroup and its child resources may not actually be stored in the location in this name.

pub fn locations_entry_groups_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationEntryGroupSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy for a resource. Replaces any existing policy. Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. Callers must have following Google IAM permission - datacatalog.tagTemplates.setIamPolicy to set policies on tag templates. - datacatalog.entries.setIamPolicy to set policies on entries. - datacatalog.entryGroups.setIamPolicy to set policies on entry groups.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_entry_groups_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationEntryGroupTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns the caller's permissions on a resource. If the resource does not exist, an empty set of permissions is returned (We don't return a NOT_FOUND error). Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. A caller is not required to have Google IAM permission to make this request.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_tag_templates_fields_enum_values_rename(
&self,
request: GoogleCloudDatacatalogV1beta1RenameTagTemplateFieldEnumValueRequest,
name: &str
) -> ProjectLocationTagTemplateFieldEnumValueRenameCall<'a, C>

Create a builder to help you perform the following task:

Renames an enum value in a tag template. The enum values have to be unique within one enum field. Thus, an enum value cannot be renamed with a name used in any other enum value within the same enum field.

Arguments

request - No description provided.

name - Required. The name of the enum field value. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id}/fields/{tag_template_field_id}/enumValues/{enum_value_display_name}

pub fn locations_tag_templates_fields_create(
&self,
request: GoogleCloudDatacatalogV1beta1TagTemplateField,
parent: &str
) -> ProjectLocationTagTemplateFieldCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a field in a tag template. The user should enable the Data Catalog API in the project identified by the parent parameter (see Data Catalog Resource Project for more information).

Arguments

request - No description provided.

parent - Required. The name of the project and the template location region. Example: * projects/{project_id}/locations/us-central1/tagTemplates/{tag_template_id}

pub fn locations_tag_templates_fields_delete(
&self,
name: &str
) -> ProjectLocationTagTemplateFieldDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a field in a tag template and all uses of that field. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

name - Required. The name of the tag template field to delete. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id}/fields/{tag_template_field_id}

pub fn locations_tag_templates_fields_patch(
&self,
request: GoogleCloudDatacatalogV1beta1TagTemplateField,
name: &str
) -> ProjectLocationTagTemplateFieldPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a field in a tag template. This method cannot be used to update the field type. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

request - No description provided.

name - Required. The name of the tag template field. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id}/fields/{tag_template_field_id}

pub fn locations_tag_templates_fields_rename(
&self,
request: GoogleCloudDatacatalogV1beta1RenameTagTemplateFieldRequest,
name: &str
) -> ProjectLocationTagTemplateFieldRenameCall<'a, C>

Create a builder to help you perform the following task:

Renames a field in a tag template. The user should enable the Data Catalog API in the project identified by the name parameter (see Data Catalog Resource Project for more information).

Arguments

request - No description provided.

name - Required. The name of the tag template. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id}/fields/{tag_template_field_id}

pub fn locations_tag_templates_create(
&self,
request: GoogleCloudDatacatalogV1beta1TagTemplate,
parent: &str
) -> ProjectLocationTagTemplateCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a tag template. The user should enable the Data Catalog API in the project identified by the parent parameter (see Data Catalog Resource Project for more information).

Arguments

request - No description provided.

parent - Required. The name of the project and the template location [region](https://cloud.google.com/data-catalog/docs/concepts/regions. Example: * projects/{project_id}/locations/us-central1

pub fn locations_tag_templates_delete(
&self,
name: &str
) -> ProjectLocationTagTemplateDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a tag template and all tags using the template. Users should enable the Data Catalog API in the project identified by the name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

name - Required. The name of the tag template to delete. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id}

pub fn locations_tag_templates_get(
&self,
name: &str
) -> ProjectLocationTagTemplateGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a tag template.

Arguments

name - Required. The name of the tag template. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id}

pub fn locations_tag_templates_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationTagTemplateGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the access control policy for a resource. A NOT_FOUND error is returned if the resource does not exist. An empty policy is returned if the resource exists but does not have a policy set on it. Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. Callers must have following Google IAM permission - datacatalog.tagTemplates.getIamPolicy to get policies on tag templates. - datacatalog.entries.getIamPolicy to get policies on entries. - datacatalog.entryGroups.getIamPolicy to get policies on entry groups.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_tag_templates_patch(
&self,
request: GoogleCloudDatacatalogV1beta1TagTemplate,
name: &str
) -> ProjectLocationTagTemplatePatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a tag template. This method cannot be used to update the fields of a template. The tag template fields are represented as separate resources and should be updated using their own create/update/delete methods. Users should enable the Data Catalog API in the project identified by the tag_template.name parameter (see [Data Catalog Resource Project] (https://cloud.google.com/data-catalog/docs/concepts/resource-project) for more information).

Arguments

request - No description provided.

name - The resource name of the tag template in URL format. Example: * projects/{project_id}/locations/{location}/tagTemplates/{tag_template_id} Note that this TagTemplate and its child resources may not actually be stored in the location in this name.

pub fn locations_tag_templates_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationTagTemplateSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the access control policy for a resource. Replaces any existing policy. Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. Callers must have following Google IAM permission - datacatalog.tagTemplates.setIamPolicy to set policies on tag templates. - datacatalog.entries.setIamPolicy to set policies on entries. - datacatalog.entryGroups.setIamPolicy to set policies on entry groups.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_tag_templates_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationTagTemplateTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns the caller's permissions on a resource. If the resource does not exist, an empty set of permissions is returned (We don't return a NOT_FOUND error). Supported resources are: - Tag templates. - Entries. - Entry groups. Note, this method cannot be used to manage policies for BigQuery, Pub/Sub and any external Google Cloud Platform resources synced to Data Catalog. A caller is not required to have Google IAM permission to make this request.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_taxonomies_policy_tags_create(
&self,
request: GoogleCloudDatacatalogV1beta1PolicyTag,
parent: &str
) -> ProjectLocationTaxonomyPolicyTagCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a policy tag in the specified taxonomy.

Arguments

request - No description provided.

parent - Required. Resource name of the taxonomy that the policy tag will belong to.

pub fn locations_taxonomies_policy_tags_delete(
&self,
name: &str
) -> ProjectLocationTaxonomyPolicyTagDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a policy tag. Also deletes all of its descendant policy tags.

Arguments

name - Required. Resource name of the policy tag to be deleted. All of its descendant policy tags will also be deleted.

pub fn locations_taxonomies_policy_tags_get(
&self,
name: &str
) -> ProjectLocationTaxonomyPolicyTagGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a policy tag.

Arguments

name - Required. Resource name of the requested policy tag.

pub fn locations_taxonomies_policy_tags_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationTaxonomyPolicyTagGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the IAM policy for a taxonomy or a policy tag.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_taxonomies_policy_tags_list(
&self,
parent: &str
) -> ProjectLocationTaxonomyPolicyTagListCall<'a, C>

Create a builder to help you perform the following task:

Lists all policy tags in a taxonomy.

Arguments

parent - Required. Resource name of the taxonomy to list the policy tags of.

pub fn locations_taxonomies_policy_tags_patch(
&self,
request: GoogleCloudDatacatalogV1beta1PolicyTag,
name: &str
) -> ProjectLocationTaxonomyPolicyTagPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a policy tag.

Arguments

request - No description provided.

name - Output only. Resource name of this policy tag, whose format is: "projects/{project_number}/locations/{location_id}/taxonomies/{taxonomy_id}/policyTags/{id}".

pub fn locations_taxonomies_policy_tags_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationTaxonomyPolicyTagSetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the IAM policy for a taxonomy or a policy tag.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_taxonomies_policy_tags_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationTaxonomyPolicyTagTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns the permissions that a caller has on the specified taxonomy or policy tag.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_taxonomies_create(
&self,
request: GoogleCloudDatacatalogV1beta1Taxonomy,
parent: &str
) -> ProjectLocationTaxonomyCreateCall<'a, C>

Create a builder to help you perform the following task:

Creates a taxonomy in the specified project.

Arguments

request - No description provided.

parent - Required. Resource name of the project that the taxonomy will belong to.

pub fn locations_taxonomies_delete(
&self,
name: &str
) -> ProjectLocationTaxonomyDeleteCall<'a, C>

Create a builder to help you perform the following task:

Deletes a taxonomy. This operation will also delete all policy tags in this taxonomy along with their associated policies.

Arguments

name - Required. Resource name of the taxonomy to be deleted. All policy tags in this taxonomy will also be deleted.

pub fn locations_taxonomies_export(
&self,
parent: &str
) -> ProjectLocationTaxonomyExportCall<'a, C>

Create a builder to help you perform the following task:

Exports all taxonomies and their policy tags in a project. This method generates SerializedTaxonomy protos with nested policy tags that can be used as an input for future ImportTaxonomies calls.

Arguments

parent - Required. Resource name of the project that taxonomies to be exported will share.

pub fn locations_taxonomies_get(
&self,
name: &str
) -> ProjectLocationTaxonomyGetCall<'a, C>

Create a builder to help you perform the following task:

Gets a taxonomy.

Arguments

name - Required. Resource name of the requested taxonomy.

pub fn locations_taxonomies_get_iam_policy(
&self,
request: GetIamPolicyRequest,
resource: &str
) -> ProjectLocationTaxonomyGetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Gets the IAM policy for a taxonomy or a policy tag.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being requested. See the operation documentation for the appropriate value for this field.

pub fn locations_taxonomies_import(
&self,
request: GoogleCloudDatacatalogV1beta1ImportTaxonomiesRequest,
parent: &str
) -> ProjectLocationTaxonomyImportCall<'a, C>

Create a builder to help you perform the following task:

Imports all taxonomies and their policy tags to a project as new taxonomies. This method provides a bulk taxonomy / policy tag creation using nested proto structure.

Arguments

request - No description provided.

parent - Required. Resource name of project that the imported taxonomies will belong to.

pub fn locations_taxonomies_list(
&self,
parent: &str
) -> ProjectLocationTaxonomyListCall<'a, C>

Create a builder to help you perform the following task:

Lists all taxonomies in a project in a particular location that the caller has permission to view.

Arguments

parent - Required. Resource name of the project to list the taxonomies of.

pub fn locations_taxonomies_patch(
&self,
request: GoogleCloudDatacatalogV1beta1Taxonomy,
name: &str
) -> ProjectLocationTaxonomyPatchCall<'a, C>

Create a builder to help you perform the following task:

Updates a taxonomy.

Arguments

request - No description provided.

name - Output only. Resource name of this taxonomy, whose format is: "projects/{project_number}/locations/{location_id}/taxonomies/{id}".

pub fn locations_taxonomies_set_iam_policy(
&self,
request: SetIamPolicyRequest,
resource: &str
) -> ProjectLocationTaxonomySetIamPolicyCall<'a, C>

Create a builder to help you perform the following task:

Sets the IAM policy for a taxonomy or a policy tag.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy is being specified. See the operation documentation for the appropriate value for this field.

pub fn locations_taxonomies_test_iam_permissions(
&self,
request: TestIamPermissionsRequest,
resource: &str
) -> ProjectLocationTaxonomyTestIamPermissionCall<'a, C>

Create a builder to help you perform the following task:

Returns the permissions that a caller has on the specified taxonomy or policy tag.

Arguments

request - No description provided.

resource - REQUIRED: The resource for which the policy detail is being requested. See the operation documentation for the appropriate value for this field.